% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dictionary.R
\name{apply_dictionary}
\alias{apply_dictionary}
\title{Applies a dictionary file}
\usage{
apply_dictionary(data, dictionary, variable_name, variable_label, value_labels)
}
\arguments{
\item{data}{input dataset}

\item{dictionary}{dictionary dataset}

\item{variable_name}{dictionary field for variable names in dataset}

\item{variable_label}{dictionary field for variable labels}

\item{value_labels}{dictionary field for pipe delimited value labels (e.g. 1, Yes | 0, No)}
}
\value{
A dataset with variable labels and factor labels
}
\description{
Applies a dictionary file
}
\examples{
apply_dictionary(data = trees,
                 dictionary = trees_dictionary,
                 variable_name = var_name,
                 variable_label = var_label,
                 value_labels = value_label)

}
