% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_quarto_tbl.R
\name{as_quarto_tbl}
\alias{as_quarto_tbl}
\title{Formatting gtsummary table objects for quarto}
\usage{
as_quarto_tbl(data)
}
\arguments{
\item{data}{A gtsummary table Object}
}
\value{
A kable table object
}
\description{
Formatting gtsummary table objects for quarto
}
\examples{
table1 <-
  trees \%>\%
  gtsummary::tbl_summary(include = c(height,diameter))

as_quarto_tbl(table1)
}
