% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age}
\usage{
calculate_age(dob, date)
}
\arguments{
\item{dob}{Date of birth}

\item{date}{Date for which age is calculated}
}
\value{
Returns an integer
}
\description{
Calculate age
}
\examples{
calculate_age(dob = "1980-10-01", date = "2024-01-01")
}
