% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_months.R
\name{calculate_months}
\alias{calculate_months}
\title{Calculate number of months between two dates with boundaries at the day of the month.}
\usage{
calculate_months(start, end)
}
\arguments{
\item{start}{Start date}

\item{end}{End date}
}
\value{
Returns an integer
}
\description{
Calculate number of months between two dates with boundaries at the day of the month.
}
\examples{
calculate_months(start = "2022-10-01", end = "2024-01-01")
}
