% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_to_integer.R
\name{char_to_integer}
\alias{char_to_integer}
\title{Convert any character columns containing integers to integer types}
\usage{
char_to_integer(data)
}
\arguments{
\item{data}{Dataset}
}
\value{
Dataset with updated data types
}
\description{
Convert any character columns containing integers to integer types
}
\examples{
char_to_integer(data = mtcars)
}
