% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/megafun-package.R
\docType{package}
\name{megafun-package}
\alias{megafun}
\alias{megafun-package}
\title{megafun: Collection of functions for analysis and reporting}
\description{
Collection of functions for simplifying data ceaning, analysis, and generating reports.
}
\author{
\strong{Maintainer}: Megan Holtorf \email{megan.holtorf@gmail.com}

}
\keyword{internal}
