% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trees_dictionary}
\alias{trees_dictionary}
\title{Data dictionary for "trees" dataset}
\format{
A data frame with 23 rows and 3 variables:
\describe{
\item{var_name}{Variable Nammes}
\item{var_label}{Variable labels}
\item{value_label}{Values and labels}
}
}
\usage{
trees_dictionary
}
\description{
Data dictionary for "trees" dataset
}
\keyword{datasets}
